Change wasmparser's license to "Apache-2.0 WITH LLVM-exception".
We are planning to make a minor change to wasmparser's license, changing it from "Apache-2.0" to "Apache-2.0 WITH LLVM-exception". This is the license the LLVM project is planning to transition to [0], and it is registered with SPDX [1].

[0] http://llvm.org/foundation/relicensing/
[1] https://spdx.org/licenses/LLVM-exception.html

LLVM-exception adds a clause providing that the output of compilation isn't subject to the attribution requirements, and a clause providing explicit GPLv2 compatibility. The full text of the license can be found on the LLVM relicensing page [2].

[2] http://llvm.org/foundation/relicensing/LICENSE.txt