Add experimental support for bulk memory operations
This commit adds support for the experimental bulk memory operations
proposal at https://github.com/WebAssembly/bulk-memory-operations. This
implements the proposal largely to the best extent we can, but in
particular passive element segments remain somewhat questionable because
their initializers are supposed to be opcodes but the encodings there
haven't been settled with the upstream reference types proposal just
yet.