Move wasm-opt-ttf to a standalone test
The failure output on CI was inscrutable and couldn't be reproduced.
Instead let's have a dedicated test which ideally has better output on
failure. In the process learn that `wasm-opt` uses `wabt`'s encoding of
`if` which doesn't quite seem to be standard, and this should fix the
tests to run `wasm-opt` again.