Expose results of name resolution (#30)
This commit exposes the `Module::resolve` method which allows accessing
the results of name resolution. The thinking here is that formats like
wasm interface types will want to access the function indices of the
core module but by name as well, so we can expose the results of name
resolution so the interface types section can access the same results.