Implement the GC proposal (#67)
* Add 'gc_feature_optin' module field

This commit adds a (gc_feature_optin $version:u8) module field.

The field is encoded as a standalone section before all other sections,
with ID=42 and contains a single byte for the version of the GC proposal
the user is opting into.

* Implement (ref T) and (struct (field valtype)..)

This commit extends value type to contain (ref T), and type to define
structs.

There are two large and slightly entangled changes.
 (1) ValType::Ref(Index) which requires ValType's to have their names
     resolved.
 (2) Type contains TypeDef which may be FunctionType or StructType.

The registration pass of resolve::names is extended to register field
names which are globally unique, but resolve to a relative index to the
struct they are defined on.

Every Ast item that contains a ValType is updated to resolve the
potential (ref T) they contain. This includes Types, Globals, Funcs,
Select, Block, If, Loop, CallIndirect.

TypeUse is in a bit of an ambiguous place as it's only used for
referencing/declaring a function type. There is no way to define a
struct type inline. The name might need to be updated down the road, but
that should probably wait for an actual GC spec to be developed.

The remaining GC proposal instructions are also implemented and are
fairly straightforward.

* Add basic parse tests for GC

This will give the 'wat' repo some basic testing of the GC proposal
while wabt is implementing it.