Remove funcref as a subtype of anyref (#85)
* Remove `nullref` type

* Refactor RefType out of ValType

No functional changes. This will help parsing of type-immediate
ref.null and ref.is_null. Eventually we may be able to combine
TableElemType with RefType as well.

* Make ref.null/is_null use a ref-type immediate

The type uses the formal reference-types production, so it's `ref.null
any` not `ref.null anyref`.

The elem segment parser also had to be updated to check the type of
ref.null used is compatible with the element segment type. This is
tricky with subtypes, so for now it just does strict equality.

* Rename `anyref` to `externref` but leave a backwards-compat alias

`anyref` is going to be a tricky thing to rename, if even possible. For
now an alias seems to be reasonable.

* Fix regression/gc-struct to use ref-type immediate ref.null

* Mark a bunch of wabt tests as failing until they update

This is unfortunate, but I don't see a way around it if we want to not
be blocked on wabt. At the least, we do have extensive test coverage in
spidermonkey for these testsuites that will continue.

* Respond to review feedback