Update to latest model of RefType
The function-references and gc proposals now model reftype as:
  heaptype := func, extern, exn, <index>
  reftype := ref null? <heaptype>

And retro-defines externref, funcref, etc as abbreviations to this core
form.

This commit splits HeapType from RefType and adds a 'nullable' flag. The
encodings are also updated to match. This commit does not expand parsing
to cover the new forms that are allowed by this model.