[wat] Update reftype parsing to match latest proposals (#31)
* Update to latest model of RefType

The function-references and gc proposals now model reftype as:
  heaptype := func, extern, exn, <index>
  reftype := ref null? <heaptype>

And retro-defines externref, funcref, etc as abbreviations to this core
form.

This commit splits HeapType from RefType and adds a 'nullable' flag. The
encodings are also updated to match. This commit does not expand parsing
to cover the new forms that are allowed by this model.

* Update parsing of RefType

This commit updates the parsing of reftype to match the latest
proposals.

Notably, (optref) is gone, (ref T) is non-nullable, and nullability is
given by a 'null' symbol not an 'opt' symbol.

* Drop anyref legacy support

SpiderMonkey has removed all uses of this syntax and no-longer needs it.

* Generalize TableElemType to RefType

* Refine ref.null to take HeapType

* Fix doc example using TableElemType