wasmpaser: Validate nested modules (#30)
* wasmpaser: Validate nested modules

This commit implements recursive validation of nested modules in
wasmparser. Previously nested modules were simply skipped but now
they're recursed into and actually checked. This commit also comes with
a more complete implementation of handling `alias` directives.

Internally this required quite a bit of refactoring. The validator now
retains a stack of modules which are being validated and remembers
parent-relationships between them. Indices into this stack are then used
for recording type definitions. The type of a
function/instance/module/etc can be defined the current module or any
previous module on the stack. New helper functiosn were added to help
resolve this new `Def` type to ensure it's handled correctly.

* Fix top-level `validate` function

This was previously trying to validate all functions in the context of
the top-level module, when instead it needed to validate each function
within the right module.