wast: Factor out StorageType from ValType
The GC proposal uses a storagetype abstraction to model what can be
stored in a GC object. This is any valuetype along with two packed
integer types, i8 and i16. Originally these were in ValueType, which
allows them to be parsed in many places. Factoring them out to
StorageType makes parsing more spec-compliant.