Implement function-references proposal and update GC proposal (#56)
* wast: Factor out StorageType from ValType

The GC proposal uses a storagetype abstraction to model what can be
stored in a GC object. This is any valuetype along with two packed
integer types, i8 and i16. Originally these were in ValueType, which
allows them to be parsed in many places. Factoring them out to
StorageType makes parsing more spec-compliant.

* wast: Add depth to rtt

The latest version of the GC proposal has a depth immediate to the 'rtt'
value type.

* wast: Update GC proposal instructions

This commit updates the GC proposal instructions to match the latest
development upstream and in V8. The canonical version I'm tracking is in
the V8 prototype document [1], although the spec is aligned with that.

[1] https://docs.google.com/document/d/1DklC3qVuOdLHSXB5UXghM_syCh-4cMinQ50ICiXnK3Q/edit

* wast: Implement `let`

The final instruction missing from function-references/gc is the 'let'
instruction.

'let' is a block instruction just like 'block','loop','if', with the
addition of introducing new locals taken from the stack.

Examples:
```
(let (result i32) (local i32 i32) <instructions>)

let (param i32) (local i32)
  <instructions>
end
```

 * Local parsing and encoding was factored out of `Func` to be re-used by `Let`
 * Name resolution of expressions was expanded to push/pop a depth of
   locals as new blocks are entered/left. Let blocks push new locals.

* wast: Make heaptype on ref.null inside assert_return optional

The interpreter allows the type annotation to be optional here.

* wast: Parse typeuse immediate in func.bind instead of function index

* wast: Fix parsing of heaptype/reftype

The function-references proposal actually specifies a heaptype as:
```
heaptype ::= ...
             (type $index)
```
And reftype has an abbreviation for:
```
(ref null? $index) => (ref null? (type $index))
```

* wast: Don't parse elem as active if it starts with reftype

Element segments are the worst to parse. Passive segments can begin with
a reftype, which can begin with a lparen, which currently forces us to
parse the elem segment as active. This commit adds lookahead to
disambiguate this.

* wast: Resolve table element types

* tests: Add basic let test

* tests: Enable testsuite/function-reference tests