Implement the multi-memory proposal
This commit implements the current state of the [multi memory
proposal][repo] where the general idea is that all instructions which
touch memory now take an index of some form to indicate which memory is
being touched. Most of this was pretty simple, it's just plumbing around
the various places that indexes appear which need to be resolved,
validated, and printed.

[repo]: https://github.com/webassembly/multi-memory