Initial implementation of memory64 proposal
This is an initial implementation of the memory64 proposal for our
tools. The main change here is validation where the index type is now
dynamic depending on the memory's type, but otherwise the text/binary
changes are pretty small. I filed a few issues for clarification along
the way and tried to "do something reasonable" in the meantime.