Place a limit on the depth of recursive module parsing
The entire `wast` crate is designed with recursive descent parsing in
mind, since it's quite easy to work with at an API layer. The downside
of this is that stack overflow happens on deeply recursive syntax. While
we handle this for expressions with an explicit `Vec` and a stack, this
would require a much larger refactoring to handle this for nested
modules.

This commit places a hardcoded limit on the depth of nested modules to
prevent the parser from stack overflowing. The hope is that this won't
ever get hit in practice, but if it does we can try to rearchitect at
least the top-level module parsing to not be as recursive.