Allow generation of possibly invalid modules
This commit takes a technique the spidermonkey folks are doing which is
to use the fuzz input as the raw body of a function. Currently many
fuzzers use the raw fuzz input as the entire wasm module, but that makes
it difficult to structure a valid wasm module because of section sizes
and such. The hope is that by simply using arbitrary fuzz input for the
instruction stream we can test more edge cases in the wasm decoder.

The changes here are to, for each function, use the fuzz input to
indicate whether a valid instruction stream is generated or a random one
is used. Modules now have a flag indicating whether they're guaranteed
valid modules (only procedurally generated functions).