Start reducing memory overhead of the parser
We've got a fuzz bug on Wasmtime right now where the wasm module
generated, when converted to text and then parsed, blows the memory
limit on the fuzz infrastructure. The memory usage and performance of
`wast` hasn't ever really been too too much of a concern, so this picks
off a low-hanging-fruit which is the size of the `Source` enum.

Before this commit `Source` was 136 bytes. The way the parser works
right now is that it lexes the entire source and creates a list of
tokens to parse, storing them into a `Vec`. This commit shrinks the size
of this element from 136 bytes to 24 (and flattens `Source` into one
`Token` enum). This drastically reduces the amount of memory allocated
to store the tokens of what we're parsing.

Unfortunately the parser is still hanging on to too much memory, so I'll
be adding a few follow-up PRs as well to handle further memory
improvements.