Optimize call_valid and global_set_valid
This fixes an issue of generating modules taking a very long time with
lots of globals and lots of functions. The problem is that the
`global_set_valid` and `call_valid` filters were taking a very long time
because they kept iterating all globals/functions to see if the relevant
types were on the stack. The fix here is to precalculate information
about the module to gather globals/functions based on type first, then
use that to see if the right types are on the stack. This makes a
previous test case that took ~30s to generate a module locally (in
sanitized release mode) take a few milliseconds afterwards.