Rewrite handling of blocks in the operator validator
This commit rewrites how blocks are handled in the `OperatorValidator`
type, primarily motivated by the various bugs we've had recently around
the handling of `unreachable` instructions (#100 being the most recent).

The goal of this refactoring is to update the validator to match, almost
exactly, the [online appendix for the validation algorithm][appendix],
namely around how blocks and `unreachable` are handled. This also should
hopefully make the code a bit easier to read since the trickiness is
handled in a relatively subtle way where the "naive interpretation" ends
up being correct.

I've done some quick benchmarking and this doesn't appear to either
regress or improve performance (yay?), but this does...

Closes #100

[appendix]: https://webassembly.github.io/spec/core/appendix/algorithm.html