Speed up code section generation
Unrolls the loop that checks whether each instruction is valid in the current
context and, if it is valid, pushes it onto our options. Unrolling this loops
lets us avoid dynamic calls through function pointers and, furthermore, each
call site can be branch predicted and even have its callee inlined.

This saved us about 30% of time in the `corpus` benchmark.