wasm-smith: add some smoke unit tests
Previously, `wasm-smith` was tested by fuzzing for 5 minutes in CI, but after
the merge into the `wasm-tools` repo, it doesn't make sense to do that anymore
because (a) added CI time, and (b) OSS-fuzz will run our fuzz targets
continuously anyways.

Instead, here are a few smoke tests to ensure that we generate valid
modules. These are just intended for local development and a quick double check
in CI.