Improve handling of names in wasmprinter
This commit fixes a longstanding issue in `wasmprinter` where names
weren't necessarily printed well. Issues arose when the name section had
non-identifier characters (or were empty) or when the name section
names were duplicates of each other.

The fix here was to add more intelligent logic to `wasmprinter` to
ensure that non-printable names (or duplicates) use `@name` annotations
which allows us to pick an arbitrary name for the function which is a
valid identifier.