Support running tests without `wabt` installed. (#138)
This commit adds support for running the test suite without the `wabt`
tools installed because they're largely optional for testing anyway. We
still run quite a few tests without them! This still requires an opt-in
to acknowledge that we're running fewer tests, but the error messages is
intended to convey that.