Fix a mismatch issue in nested wasm module validation
The code for validation the module code section was modeled after the
validation of the code section itself, but that included the flawed
assumption that the modules defined by the module code section always
came at the end of the module index space. Modules, unlike functions,
can have their definitions interleaved with imports and aliases, so the
definitions we find aren't guaranteed to define any sort of contiguous
range of modules.

The fix here is relatively simple, to just keep track of a second map of
when we reach the nth code definition it maps to the ith module in the
index space.