wasm-smith: add some additional configuration options.
This PR adds some options to more tightly control the arbitrary Wasm
modules generated by `wasm-smith`. In particular, it adds minimums
alongside most of the maximums, so that we can require certain elements
to be present (e.g., a memory or at least one function); it allows the
user to omit start functions; and it allows for a limit to be placed on
maximum Wasm memory sizes, which is useful during fuzzing (to reduce
memory overhead).