Use hex for the Debug output of Float32/Float64
When visually comparing the output of failed assertions (e.g. "expected = ..., actual = ..."), it is much easier to compare float values as hex than their unsigned integer representation. Optimally we would use actual floats in the output but this is fraught since sometimes it is necessary to do a bit-by-bit comparison. Instead, a hex representation shows the bits and avoids a manual "decimal-to-hex" conversion.