Start implementing module-linking in wasm-smith
This commit starts generating module-linking-using modules in the
`wasm-smith` module generator. It's going to take some time to implement
everything so I figured it's best to do it piecemeal. First up in this
commit is the type section, which includes generation of module and
instance types. This also applied a few internal refactorings:

* The type/import sections may be generated in an interleaved fashion
  now with module linking.
* Lookups in the `types` array are now indirected through other arrays
  that record type indices of a specific kind.
* Iterators over types/imports are now indirected through side arrays
  which index into `initializers`.

Further refactorings will be necessary to account for adjustments in
the index spaces as aliases are introduced, but this seemed like a good
start!