Add initial support for fuzzing the `alias` section
This commit adds the bare-bones support o `wasm-smith` to fuzz the
alias section of a module-linking-using module. This enables the fuzzer
to generate `alias` annotations referencing instances and pull more
items into the index space.

The main bulk of this commit is refactoring the internals of `Module` to
keep track in tables the types of all items, regardless of where they
came from. This simplifies some bits and pieces here and there and
obviates a few now no-longer-necessary caches. After the refactoring a
new `AvailableAliases` type was added to keep track of alias candidates
over time which are filtered as soon as a module's maximum requirements
are met.