Add support for fuzzing the instance section (#167)
This commit adds support for fuzzing the instance section of a
module-linking-using module. A cache of `AvailableInstantiations` is
kept to keep track of what modules can be instantiated at any point in
time. This cache is then used to, if there's something in it, generate
an `instance` which can then be later used for further instantiations
and/or aliasing.