Add support for fuzzing the module/module code section (#169)
This commit adds support to wasm-smith and wasm-encoder to generate the
module and module code sections for module-linking modules. The module
section in wasm-smith is generated by using `Arbitrary for
ConfiguredModule<C>` to create a submodule, and then the type of the
submodule is copied out of the child into the parent for insertion into
the type section. Some deduplication is attempted but in general new
module/instance/func types will always be added to type the new module.