Update the module linking proposal encoding
This ended up being a much larger patch than intended, but the goal of
this commit is to sync the module linking proposal with some recent
upstream changes. The changes to the proposal itself, relative to what
was implemented here before, are:

* The module code section is now removed and modules are listed inline
  in the module section.

* Two-level imports are now interpreted as single-level imports of
  instances. This translation makes it much easier to represent the type
  of a module, test subtyping between module types, and implement the
  `instantiate` instruction.

* The `instantiate` instruction's list of arguments now have names
  associated with them. The goal is that you're providing the names
  required by the module you're instantiating, and by using names it
  enables implementations to build up an import map.

* The `alias` construct now references instance exports by string-name
  rather than by index.

Lots of the text format syntax and binary format for module linking has
been naturally updated here as well. Not having to list a module's type
explicitly and having strings as the sources of aliases was a *huge*
simplification to the text format. This does, however, do away with the
`$i.$foo` sugar in the text format as well as the export-all syntax of
`(export $foo)`. Alias sugar will come back as a follow-up commit and
should be pretty easy to implement, but this was big enough that I
didn't want to lump in all that new syntax here too (and it only affects
the text format).

Validation had a good number of changes to handle this update as well.
The main structural change is that previously you were guaranteed that
any type of any item in your module was either declared locally or in a
parent module. This is no longer guaranteed to be the case, meaning that
the `Def` type had to go (and I never liked that anyway).

Instead of `Def` the `Validator` now contains a global list of types
that all modules and submodules push onto. This means that we can
represent implicit types by pushing onto this list (e.g. the implicit
type of a module) as well as easily implement type aliasing as you're
simply copying an index pointer. The main point of trickiness is that
this list is growing as we parse modules but we also want to be
validating previous functions with a subset of the list concurrently. A
new `SnapshotList` structure was added to assist with this.

Another change with `Validator` is that sub-`Validator` instances are no
longer created. Instead all data is fed into a singular `Validator` for
each top-level wasm blob and it will internally do what all consumers
were already doing which was handling a stack of parent modules.