Implement new alias shorthand sugar for module linking
This commit Implements the sugar syntax proposed in
WebAssembly/module-linking#26. The change here was to introduce a new
`ItemRef` construct into the `wast` crate which represents a
parenthesized reference to an item, either in the current module or a
different module (e.g. optional `outer` and optional export names).
Additionally an `IndexOrRef` type was added for parsing references which
are either a bare index or an explicit reference to an item.

All existing references to items were updated to use `ItemRef` where
necessary. Additionally all locations which previously took an `Index`
now store `ItemRef` and are parsed as `IndexOrRef`.