Fix location of injection of alias annotations
This commit fixes an issue in wasm text parsing where alias sugar in the
module linking proposal injected the `alias` directtive in the wrong
location in the module, confusing the name resolver and the index
numbering to have wast/wasmparser disagree and cause issues. More
details can be found in the code comment itself for the solution
selected.