Clarify license of the `wast` crate (#211)
This commit fixes a mistake originating from 6ae3b271 where when the
crate was originally relicensed from "MIT/Apache-2.0" to "Apache-2.0
WITH LLVM-exception" I forgot to delete the corresponding license files
or update time. The `Cargo.toml` of `wast` says that it has the
"Apache-2.0 WITH LLVM-exception" license but the license files
themselves still look like "MIT/Apache-2.0". This clarifies the
situation by deleting the extraneous files to inherit the repository's
"Apache-2.0 WITH LLVM-exception" license.