Account for invalid names in roundtrip fuzzing
This fixes a bug in the roundtrip fuzzing target where an input `*.wat`
file can produce a wasm file with an "invalid" name section because the
name might be too large. This means that the roundtrip-ing of that file
would not preserve the name section, causing the previous assertion to
fail.

The fuzzer now validates that the name section, if present, is valid
before proceeding with fuzzing.