Reorder simd opcodes in `wast`
This doesn't actually change anything, it just moves things around to
match https://webassembly.github.io/simd/core/binary/instructions.html
which is likely to be the final listing. This also, formatting-wise,
lists the second opcode as a decimal integer to match the spec as well
to avoid the need to translate between hex and back.