Reorder some SIMD opcodes (#239)
* Reorder simd opcodes in `wast`

This doesn't actually change anything, it just moves things around to
match https://webassembly.github.io/simd/core/binary/instructions.html
which is likely to be the final listing. This also, formatting-wise,
lists the second opcode as a decimal integer to match the spec as well
to avoid the need to translate between hex and back.

* Update ordering of simd instructions in wasmparser

Similar to the previous commit, update instruction orderings to match
https://webassembly.github.io/simd/core/binary/instructions.html to
ideally have a consistent ordering of all the instructions at least in
the definition.