Fix an assertion failure on invalid modules
This commit fixes an issue in wasmparser where if an invalid module used
a `ref.func` instruction in the offset of the data segment it would
attempt what has since become immutable state.

The fix here is to delay the mutation to happen until after the
typechecking happens, where if the type is valid then we're guaranteed
our state is still mutable.