Continue to tweak type size calculations in `wasm-smith`
This commit updates the type size calculation since Wasmtime was getting
requests to instantiate modules with multiple thousands of memories, but
that shouldn't be valid since the maximum type size is 1000.

The fix here is to basically say that when calculating the size of an
instance or module, adding an import/export is where the `+1` happens.
This means that the size of an instance/module is the number of
imports/exports plus the recursive size of each import/export. This fix
is applied to generating instance/module types, but it's also applied
when adding imports/exports (the `+1` happens on the addition of the
import/export.

This also refactors the loop that adds exports to build a list of all
candidate exports and then it's filtered down to only exports whose type
size fits within the remaining budget of type size.