Fix wasm-smith generating imports that exceed the type size budget
This commit fixes an issue where it's possible for imports of a module
to be generated such that they exceed the type size budget. The fix is
to generate a type to first and test if it's too large before we
actually add it. Previously items were added and then if the budget was
exceeded we'd stop adding items.