Add an "exhaustive checking mode" to `wast`
This commit adds a conditional compilation mode to `wast` where new
fields are injected into some AST nodes to add assertions later on that
all AST nodes were visited by the resolution passes. There have been
quite a few bugs recently related to the alias pass forgetting to run on
some AST nodes, so this is intended to fix the last of those once and
for all.

When the `--cfg wast_check_exhaustive` flag is used when compiling `wast`
(must be passed via `RUSTFLAGS`) the crate will insert boolean fields
into all item references along with a "visited" flag. This "visited"
flag is then set during alias expansion (and perhaps later we could
check it during resolve expansion too). Resolution and binary encoding
passes which already visit all of these nodes anyway then check this
flag and return a bug or assert if something is encountered.

Another possible solution to this would be some sort of `Visit` trait
but that sort of feels like overkill when just two passes are in
question here (alias expansion and resolution). I'm hoping that we can
leverage this to be a bit lighter weight in the long run, but we'll see!