Align type size calculation in wasmparser/wasm-smith
This commit updates the type size calculation in wasmparser to more
closely align with that of wasm-smith. Previously wasmparser would
conservatively say that the type size of the instance of an instantiated
module was the same as that of the module that was being instantiated,
but wasm-smith only considered the size of the exports. Additionally the
`+1` for each item happened in wasm-smith outside of the `size`
calculation rather than inside.

This fixes both of these issues in wasmparser's validator, the main fix
being having instances count far less towards the type size budget
because the imports of the instantiated module are no longer considered
in the type size of the instance itself.