Enable larger instruction sequences in functions
Previously, there was a 50% chance that the instruction sequence would
finish early (i.e., before we run out of random data). This probability
is reduced to 1/256, making it much more likely to create longer
instruction sequences.

Longer instruction sequences in turn seemed to frequently expose an "out
of random data" error. This change makes the more permissive choice to
simply avoid adding new instructions when we run out of data instead of
returning an error.