Add configuration options to `wasm-smith` CLI tool
This commit moves the `wasm-smith` CLI tool to the top-level with the
other tools in this repository, and then enhances it with CLI options
for all the configuration variables for a wasm module. Additionally a
`-c` json configuration file is now supported as well for commonly
shared options and less giant CLI command lines.

Additionally a flag has been added to generate a possibly-invalid
module, exposing the `MaybeInvalidModule` type from the CLI which can
help fuzz some more as well.