Allow more invalid AST states in elem expressions
Previously we would store elem segment expressions as actually optional
indices, but this meant that some parse-time errors happened instead of
later validation-time errors. Looks like the upstream test suite expects
to be able to parse some more element segments than we support (for them
to later fail in validation), so this updates the parser to parse full
expressions and store those so we can faithfully carry over those items
in to the output to get errors during wasm validation.

Closes #294