Sync with the upstream memory64 proposal
This commit updates the support for the memory64 proposal,
printing/parsing/validation, to match the latest upstream spec. I've
double-checked all the type-checking and such, and the only piece that
this crate was missing was parsing 64-bit offsets in `memarg`
structures. I've threaded through various bits and pieces to parse
32-bit offsets primarily but optionally allow 64-bit offsets when the
memory64 proposal is enabled.

I've also done some updates to the test suite so we can run all the
`*.wast` tests (or at least most) in the upstream wasm64 proposal.